Update .gitignore and downgrade styled-components version
The .gitignore file has been updated to make sure yarn-error.log is not tracked anymore as it is a file generated when yarn encounters an error. This is not relevant to the project's codebase itself and should therefore not be tracked.

In the 'package.json' file, the 'styled-components' package version has been downgraded from "^5.3.6" to "^5.3.3". This change was necessary due to incompatibility issues between the newer version and other packages in the project.